Shift trimTrailingSemiColons and applyAutoLimit
