+++ apps/studio/components/interfaces/Explorer/QueryEditor.tsx
@@ -21,9 +21,9 @@ import { DisplaySettingsButton } from './QueryCell/DisplaySettingsButton'
 import { QueryResultChart } from './QueryCell/QueryResultChart'
 import { QueryResultTable } from './QueryResultTable'
 import { type QueryDisplay, type QueryResult } from './types'
-import { applyAutoLimit } from '@/components/interfaces/SQLEditor/SQLEditor.utils'
 import { CodeEditor } from '@/components/ui/CodeEditor/CodeEditor'
 import { useExecuteSqlMutation } from '@/data/sql/execute-sql-mutation'
+import { applyAutoLimit } from '@/data/sql/utils'
 import { useLatest } from '@/hooks/misc/useLatest'
 import { useSelectedProjectQuery } from '@/hooks/misc/useSelectedProject'
